pyproject.toml
[tool.poetry]
name = "konamon"
version = "0.1.0"
description = ""
authors = ["matonix <automatonix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
plotly = "^5.24.1"
pandas = "^2.2.3"
pyarrow = "^18.0.0"
nbformat = "^5.10.4"
pydantic = "^2.9.2"
dash = "^2.18.2"
dash-daq = "^0.5.0"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
polars = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
